[Deno] Rename some files, fix imports
