+++ mod.ts
@@ -1,0 +1,4 @@
+import P from './src/poisson-disk-sampling.ts';
+export { default as default } from './src/poisson-disk-sampling.ts';
+
+console.log(P);
